refactor: Move SignupLogic to UserManager
